Refactor PerformanceUILayout for DTXManiaNX: Enhance layout configuration, add lane types, and improve asset management for UI components
