+++ DTXMania.Game/Lib/Stage/Performance/NoteRenderer.cs
@@ -5,6 +5,7 @@ using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using DTXMania.Game.Lib.Resources;
 using DTXMania.Game.Lib.Song.Components;
+using DTXMania.Game.Lib.UI.Layout;
 
 namespace DTXMania.Game.Lib.Stage.Performance
 {
